% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,FCS-method}
\title{Exportation of FCS objects}
\usage{
export(object, filename)

\S4method{export}{FCS}(object, filename = "FCS.fcs")
}
\arguments{
\item{object}{a FCS object}

\item{filename}{a character indicating the location of the FCS file to save}
}
\value{
None
}
\description{
Exports a FCS object into a FCS file.
}
