% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{[,FCS,ANY-method}
\title{Extraction of subsets of data from FCS objects}
\usage{
\S4method{[}{FCS,ANY}(x, i, j)
}
\arguments{
\item{x}{a FCS object}

\item{i}{a numeric, logical or character vector}

\item{j}{a numeric, logical or character vector}
}
\value{
a S4 object of class FCS
}
\description{
Extracts a subset of data from a FCS object.
}
\details{
For The parameter i represents a vector of cells to extract and the parameter j represents a vector of markers to extract.
}
