% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.FCS}
\alias{import.FCS}
\title{Importation of cytometry profiles from one or several FCS files}
\usage{
import.FCS(path, exclude = NULL, trans = "arcsinh",
  trans.para = switch(trans, arcsinh = list(arcsinh.scale = 5), log =
  list(log.shift = "auto", log.base = 10), none = NULL),
  trans.exclude = "cluster")
}
\arguments{
\item{path}{a character vector indicating the location to a FCS file or to a set of FCS files}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}

\item{trans}{a character specifying the name of a transformation function to apply on the marker expression intensities. Possible functions are "arcsinh" for arc sin hyperbolic transformation (default), "log" for logarithmic transformation, or "none" for no transformation}

\item{trans.para}{a named list containing parameters for the transformation. Please refer to the details section for more details}

\item{trans.exclude}{a character vector containing the marker names for which no transformation must be applied on (including the rescaling transformation)}
}
\value{
a S4 object of class FCS
}
\description{
Imports one or several cytometry profiles from a FCS file or from a set of files into a FCS object.
}
\details{
If a set of files is specified, then the files are merged in the import procedure.

Several transformations can be applied on  marker expressions via the 'trans' parameter:

The transformation functions can be parametrized using the named list 'trans.para'. The scale (cofactor) of the arcsinh transformation function can be parametrized using the 'arcsinh.scale' value. The shift of the log transformation function can be parametrized using the 'log.shift' value and the base of the log transformation function can be parametrized using the 'log.base' value. The 'log.shift' allows the value "auto" which automatically identify the log shift avoiding to apply log transformations on negative values.
}
