% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.SPADE}
\alias{import.SPADE}
\title{Importation of cytometry profiles from a SPADE analysis}
\usage{
import.SPADE(path, exclude = c("density", "cluster"))
}
\arguments{
\item{path}{a character vector indicating the location to a FCS file or to a set of FCS files}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}
}
\value{
a S4 object of class FCS
}
\description{
Imports one cytometry profiles from a SPADE analysis into a FCS object.
}
\details{
If a set of files is specified, then the files are merged in the import procedure.
}
