% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeQuantile}
\alias{normalizeQuantile}
\title{title Normalization of CELL objects}
\usage{
normalizeQuantile(list, BBmarkers)
}
\arguments{
\item{list}{a list of CELL objects to normalize}

\item{BBmarkers}{a character vector indicating the marker names on which the normalization will be applied}
}
\value{
a list of normalized CELL objects
}
\description{
Performs a normalization between two of more CELL objects.
}
