% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,FCS,missing-method}
\alias{plot,FCS,FCS-method}
\title{Plot for FCS objects}
\usage{
plot(object1, object2, ...)

\S4method{plot}{FCS,missing}(object1, object2)

\S4method{plot}{FCS,FCS}(object1, object2)
}
\arguments{
\item{object1}{a FCS to plot}

\item{object2}{a FCS to plot}

\item{...}{other parameters}
}
\value{
returns a ggplot object
}
\description{
Makes visual representations for FCS objects.
}
