% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,FCS-method}
\title{Textual preview for FCS objects}
\usage{
\S4method{print}{FCS}(x)
}
\arguments{
\item{x}{a FCS object}
}
\value{
None
}
\description{
Prints a preview for a FCS object.
}
