% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,FCS-method}
\title{Textual preview for FCS objects}
\usage{
\S4method{show}{FCS}(object)
}
\arguments{
\item{object}{a FCS object}
}
\value{
None
}
\description{
Shows a preview for a FCS object.
}
